+++ test/CubeCounting.test.js
@@ -14,6 +14,7 @@ const {
 const { evmRevert, evmSnapshot, getCallData } = require('./utils/utils');
 
 const Registry = artifacts.require('Registry');
+const FeeRuleRegistry = artifacts.require('FeeRuleRegistry');
 const Proxy = artifacts.require('Proxy');
 const HAaveV2 = artifacts.require('HAaveProtocolV2');
 const HWrapper = artifacts.require('HWmatic');
@@ -27,7 +28,11 @@ contract('CubeCounting', function([_, user]) {
 
   before(async function() {
     this.registry = await Registry.new();
-    this.proxy = await Proxy.new(this.registry.address);
+    this.feeRuleRegistry = await FeeRuleRegistry.new('0', _);
+    this.proxy = await Proxy.new(
+      this.registry.address,
+      this.feeRuleRegistry.address
+    );
 
     // Register wrapper handler
     this.hWrapper = await HWrapper.new();
@@ -65,7 +70,7 @@ contract('CubeCounting', function([_, user]) {
       const config = [ZERO_BYTES32];
       const data = [getCallData(HWrapper, 'deposit', [value])];
       await expectRevert(
-        this.proxy.batchExec(to, config, data, {
+        this.proxy.batchExec(to, config, data, [], {
           from: user,
           value: 0, // Insufficient native token
         }),
@@ -93,7 +98,7 @@ contract('CubeCounting', function([_, user]) {
       ];
 
       await expectRevert(
-        this.proxy.batchExec(to, config, data, {
+        this.proxy.batchExec(to, config, data, [], {
           from: user,
         }),
         '0_HAaveProtocolV2_flashLoan: _exec'
@@ -121,7 +126,7 @@ contract('CubeCounting', function([_, user]) {
       ];
 
       await expectRevert(
-        this.proxy.batchExec(to, config, data, {
+        this.proxy.batchExec(to, config, data, [], {
           from: user,
         }),
         '0_HAaveProtocolV2_flashLoan: SafeERC20: low-level call failed'
@@ -152,7 +157,7 @@ contract('CubeCounting', function([_, user]) {
       ];
 
       await expectRevert(
-        this.proxy.batchExec(to, config, data, {
+        this.proxy.batchExec(to, config, data, [], {
           from: user,
         }),
         '0_HAaveProtocolV2_flashLoan: 0_HWmatic_withdraw: Unspecified'
@@ -188,7 +193,7 @@ contract('CubeCounting', function([_, user]) {
       ];
 
       await expectRevert(
-        this.proxy.batchExec(to, config, data, {
+        this.proxy.batchExec(to, config, data, [], {
           from: user,
         }),
         '0_HAaveProtocolV2_flashLoan: 1_HWmatic_deposit: Unspecified'
@@ -220,7 +225,7 @@ contract('CubeCounting', function([_, user]) {
       ];
 
       await expectRevert(
-        this.proxy.batchExec(to, config, data, {
+        this.proxy.batchExec(to, config, data, [], {
           from: user,
         }),
         '1_HAaveProtocolV2_flashLoan: 0_HWmatic_withdraw: Unspecified'
@@ -242,7 +247,7 @@ contract('CubeCounting', function([_, user]) {
       const config = [ZERO_BYTES32];
       const data = [getCallData(HWrapper, 'deposit', [value])];
       await expectRevert(
-        this.proxy.batchExec(to, config, data, {
+        this.proxy.batchExec(to, config, data, [], {
           from: user,
           value: 0, // Insufficient native token
         }),