+++ test/HCurveCrypto.test.js
@@ -24,6 +24,7 @@ const {
 
 const Proxy = artifacts.require('ProxyMock');
 const Registry = artifacts.require('Registry');
+const FeeRuleRegistry = artifacts.require('FeeRuleRegistry');
 const HCurve = artifacts.require('HCurve');
 const ICurveHandler = artifacts.require('ICurveHandler');
 const IToken = artifacts.require('IERC20');
@@ -38,7 +39,11 @@ contract('Curve Crypto', function([_, user]) {
       this.hCurve.address,
       utils.asciiToHex('HCurve')
     );
-    this.proxy = await Proxy.new(this.registry.address);
+    this.feeRuleRegistry = await FeeRuleRegistry.new('0', _);
+    this.proxy = await Proxy.new(
+      this.registry.address,
+      this.feeRuleRegistry.address
+    );
     this.atricryptoDeposit = await ICurveHandler.at(CURVE_ATRICRYPTO_DEPOSIT);
 
     // FIXME: static provider beacuse curve address provider hasn't atricrypto pool